# [internal] Port `match_path_globs` to PyO3