--- a/src/rust/engine/src/externs/interface.rs
+++ b/src/rust/engine/src/externs/interface.rs
@@ -156,11 +156,6 @@ py_module_initializer!(native_engine, |py, m| {
 
   m.add(
     py,
-    "match_path_globs",
-    py_fn!(py, match_path_globs(a: PyObject, b: Vec<String>)),
-  )?;
-  m.add(
-    py,
     "write_digest",
     py_fn!(
       py,
@@ -1543,35 +1538,6 @@ fn lease_files_in_graph(
       .map(|()| None)
     })
   })
-}
-
-fn match_path_globs(
-  py: Python,
-  path_globs: PyObject,
-  paths: Vec<String>,
-) -> CPyResult<Vec<String>> {
-  let matches = py
-    .allow_threads(|| {
-      let path_globs = nodes::Snapshot::lift_prepared_path_globs(&path_globs.into())?;
-
-      Ok(
-        paths
-          .into_iter()
-          .map(PathBuf::from)
-          .filter(|pb| path_globs.matches(pb.as_ref()))
-          .collect::<Vec<_>>(),
-      )
-    })
-    .map_err(|e: String| PyErr::new::<exc::ValueError, _>(py, (e,)))?;
-
-  matches
-    .into_iter()
-    .map(|pb| {
-      pb.into_os_string().into_string().map_err(|s| {
-        PyErr::new::<exc::Exception, _>(py, (format!("Could not decode {:?} as a string.", s),))
-      })
-    })
-    .collect::<Result<Vec<_>, _>>()
 }
 
 fn capture_snapshots(