@davidhewitt 2021-06-18
I'll take a more detailed look at this on the weekend (probably Sunday). We should be able to avoid a slowdown, so I'll investigate and propose solutions.