@Eric-Arellano 2021-07-22
Great! Using PyO3 0.14.1 and Daniel's suggestion of `Path::new` instead of `PathBuf::from`, I now get PyO3 as slightly faster but essentially equivalent:

Before

```
❯ hyperfine --warmup 1 --runs=100 './pants'
  Time (mean ± σ):     698.7 ms ±   4.6 ms    [User: 584.0 ms, System: 94.4 ms]
  Range (min … max):   679.7 ms … 714.6 ms    100 runs
```

After:

```
❯ hyperfine --warmup 1 --runs=100 './pants'
  Time (mean ± σ):     697.4 ms ±   5.0 ms    [User: 581.5 ms, System: 95.4 ms]
  Range (min … max):   685.8 ms … 718.2 ms    100 runs
```